Began work on externalizing the link algorithms, in a DRY effort and working towards an uber-linking command that's so context-sensitive and artificially intelligent you'll swear your computer is reading your mind. Or something like that.
This commit includes the new library, and modifies the first "Link" command to use it. The command has an added feature now: if you have multiple links in your clipboard and you select a word, triggering the command will give you a menu of available clipboard links to apply to the word. Also works for pasting reference links, which will eventually be part of the same command.