Meta header specificity (doctype, title, link, categories, tags only).
Set tag fill to do them one at a time to avoid truncation of tags in Blogsmith, which was happening on long tag strings or strings which contained an @tag.

Figured out how to enable the save button automatically, without requiring any focus/space action on the title or content.